update to ststm 14.0

+++ Marlin/src/HAL/STM32/HAL.h
@@ -123,7 +123,7 @@
 // Types
 // ------------------------
 
-typedef int16_t pin_t;
+typedef uint32_t pin_t;
 
 #define HAL_SERVO_LIB libServo
 #define PAUSE_SERVO_OUTPUT() libServo::pause_all_servos()
